frontend/requirements.txt
streamlit
pvporcupine
sounddevice
python-dotenv
requests
audio-recorder-streamlit
